Consolidate LuceneSupport
